+++ buildSrc/src/main/kotlin/utils.kt
@@ -60,9 +60,6 @@ val publicModules = setOf(
     "sketch-view-koin",
 )
 
-val Project.lowMinSdk: Int
-    get() = intProperty("lowMinSdk")
-
 val Project.minSdk: Int
     get() = intProperty("minSdk")