+++ CHANGELOG.md
@@ -9,6 +9,12 @@ Translations: [简体中文](CHANGELOG.zh.md)
      prompt for upgrade.
 > 3. Reference [《Migration Documentation》](docs/migrate.md) migrating from 3.x to 4.x
 
+# new
+
+## Android
+
+* change: The minimum Android platform version of all modules has been upgraded to 23
+
 # 4.5.0-beta05
 
 ## ios